feat: refactor Stats component to use centralized chart configuration helper functions
